Final for example
There are many more things that could be done and other methods that
could be employed to increase the uncoupling of the message callers and
message handlers but this is a good example of what could be done.
The best idea is to use functionality provided by the operating system,
e.g. software interupts, or other uncoupling mechanisms if they are
available. But they all work somewhat the same in the sense that they
will in the end execute a predetermined callback function as a result of
some trigger like a message or an interrupt.